requirements.txt
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.0
configparser==5.2.0
cycler==0.11.0
Flask==2.1.0
fonttools==4.31.2
googletrans==4.0.0rc1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sniffio==1.2.0
textblob==0.17.1
threadpoolctl==3.1.0
tqdm==4.63.1
tweepy==4.8.0
urllib3==1.26.9
Werkzeug==2.1.0
wordcloud==1.8.1
zipp==3.7.0
